+++ client/internal/models/models.go
@@ -1,5 +1,12 @@
 package models
 
+type User struct {
+	ID             string `json:"id"`
+	Login          string `json:"login"`
+	Password       string `json:"password"`
+	HashedPassword string `json:"hashed_password"`
+}
+
 type Secret struct {
 	ID      string `json:"id"`
 	Name    string `json:"name"`
@@ -8,5 +15,3 @@ type Secret struct {
 	UserID  string `json:"user_id"`
 	Version string `json:"version"`
 }
-
-func MakeSecret() Secret { return Secret{} }
