+++ client/internal/tui/menu.go
@@ -9,7 +9,7 @@ import (
 func (t *Tui) Menu() {
 	menu := []menuItem{
 		{name: "My secrets", shortcut: 'a', target: func() { t.SecretList() }},
-		{name: "Add secret", shortcut: 'b', target: func() { t.SecretCreatePage(models.MakeSecret()) }},
+		{name: "Add secret", shortcut: 'b', target: func() { t.SecretCreatePage(models.Secret{}) }},
 		{name: "Exit", shortcut: 'c', target: func() { t.application.Stop() }},
 	}
