+++ server/internal/models/models.go
@@ -1,5 +1,12 @@
 package models
 
+type User struct {
+	ID             string `json:"id"`
+	Login          string `json:"login"`
+	Password       string `json:"password"`
+	HashedPassword string `json:"hashed_password"`
+}
+
 type Secret struct {
 	ID      string `json:"id"`
 	Name    string `json:"name"`
@@ -7,8 +14,4 @@ type Secret struct {
 	Meta    string `json:"meta"`
 	UserID  string `json:"user_id"`
 	Version string `json:"version"`
-}
-
-func MakeSecret() Secret {
-	return Secret{}
 }
